refactor(config): 统一配置层，合并 core/config 到 config/
- core/config/ 4 个 Pydantic 模型移入 config/
- app_config.py 改名为 app_meta.py（语义更清晰）
- core/config/__init__.py 的 re-export 合并到 config/__init__.py
- 全局更新 import 路径
