+++ src/danmaku_sender/config/validation_config.py
