+++ src/danmaku_sender/core/database/history_manager.py
@@ -9,7 +9,7 @@ from playhouse.migrate import SqliteMigrator, migrate
 
 from .orm_models import db, SentDanmaku
 
-from danmaku_sender.config.app_config import AppInfo
+from danmaku_sender.config.app_meta import AppInfo
 from danmaku_sender.types.models.danmaku import Danmaku
 from danmaku_sender.types.models.common import DanmakuStatus, VideoTarget
