+++ src/danmaku_sender/main.py
@@ -5,7 +5,7 @@ from platformdirs import user_data_dir
 
 from PySide6.QtWidgets import QApplication
 
-from .config.app_config import AppInfo
+from .config.app_meta import AppInfo
 from .ui.framework.style_loader import get_app_icon
 from .utils.log_utils import init_app_logging
