+++ src/danmaku_sender/ui/account_manager/page.py
@@ -11,7 +11,7 @@ from .components import AccountRow
 from .account_form import AccountFormDialog
 
 from danmaku_sender.types.models.account import AccountCredential
-from danmaku_sender.core.config import ApiAuthConfig
+from danmaku_sender.config import ApiAuthConfig
 from danmaku_sender.runtime.app_state import AppState
 from danmaku_sender.ui.controllers.account_controller import AccountController
 from danmaku_sender.ui.framework.style_loader import get_svg_icon
