+++ src/danmaku_sender/ui/controllers/auth_controller.py
@@ -4,7 +4,7 @@ import threading
 from PySide6.QtCore import QObject, Signal, Slot
 
 from danmaku_sender.api.bili_api_client import BiliApiClient
-from danmaku_sender.core.config import ApiAuthConfig
+from danmaku_sender.config import ApiAuthConfig
 from danmaku_sender.types.models.user import UserProfile
 from danmaku_sender.ui.framework.concurrency import WorkerThread, PoolTask
