+++ src/danmaku_sender/ui/dialogs.py
@@ -14,7 +14,7 @@ from PIL import Image
 from .controllers.auth_controller import AuthController
 
 from ..types.models.common import AuthCookies
-from ..config.app_config import AppInfo, Links
+from ..config.app_meta import AppInfo, Links
 from ..utils.path_utils import get_assets_path
